Removes hard-coded attachment to core/httpd
The original implementation modified the core/httpd package. It also did
not allow for the changing of configuration in the apache application.

This adds the configuration file necessary to run httpd. This also
adds the default values for most of the settings used. To make that work
an init hook that had to be added that takes care of symlinking some
directories out of the core/httpd package. And setting the right
permissions to allow the file to be seen.

Signed-off-by: Franklin Webber <franklin@chef.io>